Task

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,6 +1,6 @@
 #include "lists.h"
 #include <stdio.h>
-
+#include <stdlib.h>
 /**
  * print_list - Prints all the elements of a list_t list.
  * @h: A pointer to the head of the list.
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
